--- light-client/starknet-storage-verifier/src/converter/mod.rs
@@ -1,1 +1,0 @@
-pub mod value;
