+++ light-client/starknet-storage-verifier/src/ibc/keys.rs
@@ -1,0 +1,42 @@
+use ibc_core::host::types::identifiers::{ChannelId, ConnectionId, PortId, Sequence};
+use poseidon::Poseidon3Hasher;
+use starknet_crypto::Felt;
+
+use super::utils::{serialize_byte_array, serialize_to_felts};
+
+pub fn connection_key(connection_id: ConnectionId) -> Felt {
+    let mut raw_path: Vec<Felt> = vec![];
+    raw_path.extend(serialize_byte_array(b"connections"));
+    raw_path.extend(serialize_byte_array(connection_id.as_bytes()));
+
+    Poseidon3Hasher::digest(&raw_path)
+}
+
+pub fn next_sequence_key(prefix: &str, port_id: PortId, channel_id: ChannelId) -> Felt {
+    let mut raw_path: Vec<Felt> = vec![];
+    raw_path.extend(serialize_byte_array(prefix.as_bytes()));
+    raw_path.extend(serialize_byte_array("ports".as_bytes()));
+    raw_path.extend(serialize_byte_array(port_id.as_bytes()));
+    raw_path.extend(serialize_byte_array("channels".as_bytes()));
+    raw_path.extend(serialize_byte_array(channel_id.as_bytes()));
+
+    Poseidon3Hasher::digest(&raw_path)
+}
+
+pub fn packet_key(
+    prefix: &str,
+    port_id: PortId,
+    channel_id: ChannelId,
+    sequence: Sequence,
+) -> Felt {
+    let mut raw_path: Vec<Felt> = vec![];
+    raw_path.extend(serialize_byte_array(prefix.as_bytes()));
+    raw_path.extend(serialize_byte_array("ports".as_bytes()));
+    raw_path.extend(serialize_byte_array(port_id.as_bytes()));
+    raw_path.extend(serialize_byte_array("channels".as_bytes()));
+    raw_path.extend(serialize_byte_array(channel_id.as_bytes()));
+    raw_path.extend(serialize_byte_array("sequences".as_bytes()));
+    raw_path.extend(serialize_to_felts(sequence.to_vec().as_slice()));
+
+    Poseidon3Hasher::digest(&raw_path)
+}
