+++ light-client/starknet-storage-verifier/src/ibc/paths.rs
@@ -1,12 +1,12 @@
 use std::str::FromStr;
 
-use ibc_core::host::types::identifiers::{ChannelId, ConnectionId, PortId, Sequence};
 use ibc_core::host::types::path::Path;
-use poseidon::Poseidon3Hasher;
 use starknet_core::utils::starknet_keccak;
 use starknet_crypto::{pedersen_hash, Felt};
 
-pub fn convert_storage_value(path: &str) -> Felt {
+use super::keys::{connection_key, next_sequence_key, packet_key};
+
+pub fn to_starknet_storage_key(path: &str) -> Felt {
     let path = Path::from_str(path).unwrap();
     match path {
         Path::Connection(connection_path) => {
@@ -92,60 +92,4 @@ pub fn convert_storage_value(path: &str) -> Felt {
 
         _ => unimplemented!(),
     }
-}
-
-pub fn connection_key(connection_id: ConnectionId) -> Felt {
-    let mut raw_path: Vec<Felt> = vec![];
-    raw_path.extend(serialize_byte_array(b"connections"));
-    raw_path.extend(serialize_byte_array(connection_id.as_bytes()));
-
-    Poseidon3Hasher::digest(&raw_path)
-}
-
-pub fn next_sequence_key(prefix: &str, port_id: PortId, channel_id: ChannelId) -> Felt {
-    let mut raw_path: Vec<Felt> = vec![];
-    raw_path.extend(serialize_byte_array(prefix.as_bytes()));
-    raw_path.extend(serialize_byte_array("ports".as_bytes()));
-    raw_path.extend(serialize_byte_array(port_id.as_bytes()));
-    raw_path.extend(serialize_byte_array("channels".as_bytes()));
-    raw_path.extend(serialize_byte_array(channel_id.as_bytes()));
-
-    Poseidon3Hasher::digest(&raw_path)
-}
-
-pub fn packet_key(
-    prefix: &str,
-    port_id: PortId,
-    channel_id: ChannelId,
-    sequence: Sequence,
-) -> Felt {
-    let mut raw_path: Vec<Felt> = vec![];
-    raw_path.extend(serialize_byte_array(prefix.as_bytes()));
-    raw_path.extend(serialize_byte_array("ports".as_bytes()));
-    raw_path.extend(serialize_byte_array(port_id.as_bytes()));
-    raw_path.extend(serialize_byte_array("channels".as_bytes()));
-    raw_path.extend(serialize_byte_array(channel_id.as_bytes()));
-    raw_path.extend(serialize_byte_array("sequences".as_bytes()));
-    raw_path.extend(serialize_to_felts(sequence.to_vec().as_slice()));
-
-    Poseidon3Hasher::digest(&raw_path)
-}
-
-fn serialize_to_felts(value: impl AsRef<[u8]>) -> Vec<Felt> {
-    let bytes = value.as_ref();
-    bytes
-        .chunks(31)
-        .map(|chunk| {
-            let mut padded = [0u8; 32];
-            padded[32 - chunk.len()..].copy_from_slice(chunk);
-            Felt::from_bytes_be(&padded)
-        })
-        .collect()
-}
-
-fn serialize_byte_array(bytes: &[u8]) -> Vec<Felt> {
-    let mut result = vec![Felt::ZERO];
-    result.extend(serialize_to_felts(bytes));
-    result.extend(vec![Felt::from(bytes.len() as u64)]);
-    result
 }
