+++ light-client/starknet-storage-verifier/src/lib.rs
@@ -1,7 +1,7 @@
-mod converter;
 mod error;
+#[cfg(feature = "ibc")]
+pub mod ibc;
 mod storage;
 
-pub use converter::*;
 pub use error::*;
 pub use storage::*;
